feat: refactor S3 integration and enhance file upload handling
