+++ src/lib/s3-server.ts
@@ -1,19 +1,22 @@
+import "server-only";
 import { S3 } from "aws-sdk";
 import * as fs from "fs";
+import { env } from "./env/server";
+import { envClient } from "./env/client";
 export async function downloadFromS3(file_key: string) {
+  console.log(file_key);
   try {
     const s3 = new S3({
-      region: process.env.NEXT_PUBLIC_AWS_REGION!,
+      region: env.AWS_REGION,
       credentials: {
-        accessKeyId: process.env.NEXT_PUBLIC_S3_ACCESS_KEY_ID!,
-        secretAccessKey: process.env.NEXT_PUBLIC_S3_SECRET_ACCESS_KEY!,
+        accessKeyId: env.AWS_ACCESS_KEY_ID,
+        secretAccessKey: env.AWS_SECRET_ACCESS_KEY,
       },
     });
     const params = {
-      Bucket: process.env.NEXT_PUBLIC_S3_BUCKET_NAME!,
+      Bucket: envClient.NEXT_PUBLIC_S3_BUCKET_NAME,
       Key: file_key,
     };
-
 
     const obj = await s3.getObject(params).promise();
     const file_name = `/tmp/pdf-${Date.now()}.pdf`;
